Android: need to not truncate to int

+++ PowerTrainer_4.4/app/src/main/java/com/stac/powertrainer_44/MainActivity.java
@@ -267,7 +267,7 @@ public class MainActivity  extends BlunoLibrary {
 
         if(mStrainEnabled) {
             tvStrain.setVisibility(View.VISIBLE);
-            tvStrain.setText("Strain: " + (int)mTrainerState.getStrain() + "");
+            tvStrain.setText("Strain: " + mTrainerState.getStrain() + "");
         } else {
             tvStrain.setVisibility(View.GONE);
         }
